proper integrity check: first commit
